Fix a segfault bug
Some input strings made fmt crash. The exact requirements for crashing
are not clear and seem to differ by architecture. In any case, the
reason (assuming this fix worked and did not merely shift the bug
elsewhere) was that the string slice backing the iterator would
sometimes be overwritten. Clearly, Rust thought that the memory was
free, which it was, since the String the slice was sliced from had
indeed gone out of scope.

The solution is to store every String in a global vector so they never
go out of scope. This is poor design in that sometimess you really are
done with a String and want to reuse its memory. However, Rust is all
about resource-aware programming, and my time is limited.